backend_fastapi/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
requests==2.32.3
python-dotenv==1.0.1
pymongo==4.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
tzdata==2025.3
python-jose[cryptography]==3.3.0
